Split root_handler into domain-specific files
