+++ bot/common/root_handler.go
@@ -2,11 +2,6 @@ package common
 
 import (
 	"fmt"
-	"regexp"
-	"slices"
-	"strconv"
-	"strings"
-
 	"github.com/bugfloyd/anonymous-telegram-bot/common/i18n"
 
 	"github.com/PaulSonOfLars/gotgbot/v2"
@@ -145,823 +140,4 @@ func (r *RootHandler) runCommand(b *gotgbot.Bot, ctx *ext.Context, command inter
 	default:
 		return fmt.Errorf("unknown command: %s", command)
 	}
-}
-
-func (r *RootHandler) processUser(userRepo *UserRepository, ctx *ext.Context) (*User, error) {
-	user, err := userRepo.readUserByUserId(ctx.EffectiveUser.Id)
-	if err != nil {
-		user, err = userRepo.createUser(ctx.EffectiveUser.Id)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return user, nil
-}
-
-func (r *RootHandler) start(b *gotgbot.Bot, ctx *ext.Context) error {
-	args := ctx.Args()
-	if len(args) == 1 && args[0] == "/start" {
-		// Reset user state
-		err := r.userRepo.resetUserState(r.user.UUID)
-		if err != nil {
-			return err
-		}
-
-		_, err = b.SendMessage(ctx.EffectiveChat.Id, i18n.T(i18n.StartMessage), &gotgbot.SendMessageOpts{})
-		if err != nil {
-			return fmt.Errorf("failed to send bot info: %w", err)
-		}
-		return nil
-	}
-	if len(args) == 2 && args[0] == "/start" {
-
-		var err error
-		var receiverUser *User
-		var identity string
-
-		if strings.HasPrefix(args[1], "_") {
-			username := args[1][1:]
-			receiverUser, err = r.userRepo.readUserByUsername(username)
-		} else {
-			receiverUser, err = r.userRepo.readUserByUUID(args[1])
-		}
-
-		if err != nil || receiverUser == nil {
-			_, err = b.SendMessage(ctx.EffectiveChat.Id, "User not found! Wrong link?", &gotgbot.SendMessageOpts{})
-			if err != nil {
-				return fmt.Errorf("failed to send bot info: %w", err)
-			}
-			return nil
-		}
-
-		if receiverUser.UUID == r.user.UUID {
-			_, err = b.SendMessage(ctx.EffectiveChat.Id, "Do you really want to talk to yourself? So sad! Share your link with friends or post it on social media to get anonymous messages!", &gotgbot.SendMessageOpts{})
-			if err != nil {
-				return fmt.Errorf("failed to send bot info: %w", err)
-			}
-			return nil
-		}
-
-		// Check if they block each other
-		blockedBy := blockCheck(&r.user, receiverUser)
-		if blockedBy != None {
-			var reason string
-			var keyboard gotgbot.InlineKeyboardMarkup
-			if blockedBy == Sender {
-				reason = "You have blocked this user."
-				keyboard = gotgbot.InlineKeyboardMarkup{
-					InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-						{
-							{
-								Text:         "Unblock",
-								CallbackData: fmt.Sprintf("ub|%s|%d", receiverUser.UUID, 0),
-							},
-						},
-					},
-				}
-
-			} else if blockedBy == Receiver {
-				reason = "This user has blocked you."
-			}
-
-			_, err = ctx.EffectiveMessage.Reply(b, reason, &gotgbot.SendMessageOpts{
-				ReplyMarkup: keyboard,
-			})
-
-			if err != nil {
-				return fmt.Errorf("failed to send block message: %w", err)
-			}
-
-			return nil
-		}
-
-		// Set user state to sending
-		err = r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-			"State":       Sending,
-			"ContactUUID": receiverUser.UUID,
-		})
-		if err != nil {
-			return fmt.Errorf("failed to update user state: %w", err)
-		}
-
-		if receiverUser.Name != "" {
-			identity = receiverUser.Name
-		} else if receiverUser.Username != "" {
-			identity = receiverUser.Username
-		} else {
-			identity = receiverUser.UUID
-		}
-
-		_, err = b.SendMessage(ctx.EffectiveChat.Id, fmt.Sprintf(i18n.T(i18n.InitialSendMessagePrompt), identity), &gotgbot.SendMessageOpts{})
-		if err != nil {
-			return fmt.Errorf("failed to send bot info: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func (r *RootHandler) info(b *gotgbot.Bot, ctx *ext.Context) error {
-	_, err := b.SendMessage(ctx.EffectiveChat.Id, "Bugfloyd Anonymous bot\n\nSource code:\nhttps://github.com/bugfloyd/anonymous-telegram-bot", &gotgbot.SendMessageOpts{})
-	if err != nil {
-		return fmt.Errorf("failed to send bot info: %w", err)
-	}
-	err = r.userRepo.resetUserState(r.user.UUID)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *RootHandler) getLink(b *gotgbot.Bot, ctx *ext.Context) error {
-	var link string
-	if r.user.Username != "" {
-		usernameLink := fmt.Sprintf("https://t.me/%s?start=_%s", b.User.Username, r.user.Username)
-		uuidLink := fmt.Sprintf("https://t.me/%s?start=%s", b.User.Username, r.user.UUID)
-		link = fmt.Sprintf("%s\n\nor:\n\n%s", usernameLink, uuidLink)
-	} else {
-		link = fmt.Sprintf("https://t.me/%s?start=%s", b.User.Username, r.user.UUID)
-	}
-	_, err := ctx.EffectiveMessage.Reply(b, link, nil)
-	if err != nil {
-		return err
-	}
-	err = r.userRepo.resetUserState(r.user.UUID)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *RootHandler) processText(b *gotgbot.Bot, ctx *ext.Context) error {
-	switch r.user.State {
-	case Sending:
-		return r.sendAnonymousMessage(b, ctx)
-	case SettingUsername:
-		return r.setUsername(b, ctx)
-	default:
-		return r.sendError(b, ctx, "Unknown Command!")
-	}
-}
-
-func (r *RootHandler) sendError(b *gotgbot.Bot, ctx *ext.Context, message string) error {
-	errorMessage := fmt.Sprintf("Error: %s", message)
-	_, err := ctx.EffectiveMessage.Reply(b, errorMessage, nil)
-	if err != nil {
-		return fmt.Errorf("failed to send error message: %w", err)
-	}
-	return nil
-}
-
-func (r *RootHandler) sendAnonymousMessage(b *gotgbot.Bot, ctx *ext.Context) error {
-	receiver, err := r.userRepo.readUserByUUID(r.user.ContactUUID)
-	if err != nil {
-		return fmt.Errorf("failed to get receiver: %w", err)
-	}
-
-	// Check if they block each other
-	blockedBy := blockCheck(&r.user, receiver)
-	if blockedBy != None {
-		var reason string
-		if blockedBy == Sender {
-			reason = "You have blocked this user."
-		} else if blockedBy == Receiver {
-			reason = "This user has blocked you."
-		}
-		_, err = ctx.EffectiveMessage.Reply(b, reason, nil)
-		if err != nil {
-			return fmt.Errorf("failed to send block message: %w", err)
-		}
-
-		// Reset sender user
-		err = r.userRepo.resetUserState(r.user.UUID)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	var replyParameters *gotgbot.ReplyParameters
-	msgText := "You have a new message."
-	if r.user.ReplyMessageID != 0 {
-		replyParameters = &gotgbot.ReplyParameters{
-			MessageId:                r.user.ReplyMessageID,
-			AllowSendingWithoutReply: true,
-		}
-
-		msgText = "New reply to your message."
-	}
-
-	// React with sent emoji to senderMessageID
-	_, err = ctx.EffectiveMessage.SetReaction(b, &gotgbot.SetMessageReactionOpts{
-		Reaction: []gotgbot.ReactionType{
-			gotgbot.ReactionTypeEmoji{
-				Emoji: "🕊",
-			},
-		},
-		IsBig: false,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to react to sender's message: %w", err)
-	}
-
-	// Send the new message notification to the receiver
-	_, err = b.SendMessage(receiver.UserID, msgText, &gotgbot.SendMessageOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-				{
-					{
-						Text:         "Open Message",
-						CallbackData: fmt.Sprintf("o|%s|%d", r.user.UUID, ctx.EffectiveMessage.MessageId),
-					},
-				},
-			},
-		},
-		ReplyParameters: replyParameters,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to send message to receiver: %w", err)
-	}
-
-	// Reset sender user
-	err = r.userRepo.resetUserState(r.user.UUID)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (r *RootHandler) openCallback(b *gotgbot.Bot, ctx *ext.Context) error {
-	cb := ctx.Update.CallbackQuery
-	split := strings.Split(cb.Data, "|")
-	if len(split) != 3 {
-		return fmt.Errorf("invalid callback data: %s", cb.Data)
-	}
-	uuid := split[1]
-	sender, err := r.userRepo.readUserByUUID(uuid)
-	if err != nil {
-		return fmt.Errorf("failed to get receiver: %w", err)
-	}
-
-	// Send callback answer to telegram
-	_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-		Text: "Message opened!",
-	})
-	if err != nil {
-		return fmt.Errorf("failed to answer callback: %w", err)
-	}
-
-	senderMessageID, err := strconv.ParseInt(split[2], 10, 64)
-	if err != nil {
-		return fmt.Errorf("failed to parse message ID: %w", err)
-	}
-	var replyMessageID int64
-	if ctx.EffectiveMessage.ReplyToMessage != nil {
-		replyMessageID = ctx.EffectiveMessage.ReplyToMessage.MessageId
-	}
-
-	// Copy the sender's message to the receiver
-	_, err = b.CopyMessage(ctx.EffectiveChat.Id, sender.UserID, senderMessageID, &gotgbot.CopyMessageOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-				{
-					{
-						Text:         "Reply",
-						CallbackData: fmt.Sprintf("r|%s|%d", sender.UUID, senderMessageID),
-					},
-					{
-						Text:         "Block",
-						CallbackData: fmt.Sprintf("b|%s|%d", sender.UUID, senderMessageID),
-					},
-				},
-			},
-		},
-		ReplyParameters: &gotgbot.ReplyParameters{
-			MessageId:                replyMessageID,
-			AllowSendingWithoutReply: true,
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("failed to send message to receiver: %w", err)
-	}
-
-	// React with eyes emoji to senderMessageID
-	_, err = b.SetMessageReaction(sender.UserID, senderMessageID, &gotgbot.SetMessageReactionOpts{
-		Reaction: []gotgbot.ReactionType{
-			gotgbot.ReactionTypeEmoji{
-				Emoji: "👀",
-			},
-		},
-		IsBig: false,
-	})
-	if err != nil {
-		fmt.Println("failed to react to sender's message: %w", err)
-	}
-
-	// Delete message with "Open" button
-	_, err = cb.Message.Delete(b, &gotgbot.DeleteMessageOpts{})
-	if err != nil {
-		fmt.Println("failed to delete message: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) replyCallback(b *gotgbot.Bot, ctx *ext.Context) error {
-	cb := ctx.Update.CallbackQuery
-	split := strings.Split(cb.Data, "|")
-	if len(split) != 3 {
-		return fmt.Errorf("invalid callback data: %s", cb.Data)
-	}
-	receiverUUID := split[1]
-	messageID, err := strconv.ParseInt(split[2], 10, 64)
-	if err != nil {
-		return fmt.Errorf("failed to parse message ID: %w", err)
-	}
-
-	// Check if receiver exists
-	receiver, err := r.userRepo.readUserByUUID(receiverUUID)
-	if err != nil {
-		return fmt.Errorf("failed to get receiver: %w", err)
-	}
-
-	// Check if they block each other
-	blockedBy := blockCheck(&r.user, receiver)
-	if blockedBy != None {
-		var reason string
-		if blockedBy == Sender {
-			reason = "You have blocked this user."
-			_, _, err = cb.Message.EditReplyMarkup(b, &gotgbot.EditMessageReplyMarkupOpts{
-				ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-					InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-						{
-							{
-								Text:         "Unblock",
-								CallbackData: fmt.Sprintf("ub|%s|%d", receiverUUID, messageID),
-							},
-						},
-					},
-				},
-			})
-
-			if err != nil {
-				return fmt.Errorf("failed to update message markup: %w", err)
-			}
-
-		} else if blockedBy == Receiver {
-			reason = "This user has blocked you."
-		}
-
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text:      reason,
-			ShowAlert: true,
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-
-		return nil
-	}
-
-	// Store the message id in the user and set status to replying
-	err = r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-		"State":          Sending,
-		"ContactUUID":    receiverUUID,
-		"ReplyMessageID": messageID,
-	})
-	if err != nil {
-		return fmt.Errorf("failed to update user state: %w", err)
-	}
-
-	// Send callback answer to telegram
-	_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-		Text: "Replying to message...",
-	})
-	if err != nil {
-		return fmt.Errorf("failed to answer callback: %w", err)
-	}
-
-	// Send reply instruction
-	_, err = ctx.EffectiveMessage.Reply(b, "Reply to this message:", nil)
-	if err != nil {
-		return fmt.Errorf("failed to send reply message: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) blockCallback(b *gotgbot.Bot, ctx *ext.Context) error {
-	cb := ctx.Update.CallbackQuery
-	split := strings.Split(cb.Data, "|")
-	if len(split) != 3 {
-		return fmt.Errorf("invalid callback data: %s", cb.Data)
-	}
-	receiverUUID := split[1]
-	replyMessageID := split[2]
-
-	err := r.userRepo.updateBlacklist(r.user.UUID, "add", receiverUUID)
-
-	if err != nil {
-		return fmt.Errorf("failed to block user: %w", err)
-	}
-
-	_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-		Text: "User blocked!",
-	})
-	if err != nil {
-		return fmt.Errorf("failed to answer callback: %w", err)
-	}
-
-	_, _, err = cb.Message.EditReplyMarkup(b, &gotgbot.EditMessageReplyMarkupOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-				{
-					{
-						Text:         "Unblock",
-						CallbackData: fmt.Sprintf("ub|%s|%s", receiverUUID, replyMessageID),
-					},
-				},
-			},
-		},
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to update message markup: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) unBlockCallback(b *gotgbot.Bot, ctx *ext.Context) error {
-	cb := ctx.Update.CallbackQuery
-	split := strings.Split(cb.Data, "|")
-	if len(split) != 3 {
-		return fmt.Errorf("invalid callback data: %s", cb.Data)
-	}
-	receiverUUID := split[1]
-	replyMessageID := split[2]
-
-	err := r.userRepo.updateBlacklist(r.user.UUID, "delete", receiverUUID)
-
-	if err != nil {
-		return fmt.Errorf("failed to unblock user: %w", err)
-	}
-
-	_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-		Text: "User unblocked!",
-	})
-	if err != nil {
-		return fmt.Errorf("failed to answer callback: %w", err)
-	}
-
-	var replyMessageKey gotgbot.InlineKeyboardButton
-	if replyMessageID == "0" {
-		replyMessageKey = gotgbot.InlineKeyboardButton{
-			Text:         "Send Message",
-			CallbackData: fmt.Sprintf("r|%s|%s", receiverUUID, replyMessageID),
-		}
-	} else {
-		replyMessageKey = gotgbot.InlineKeyboardButton{
-			Text:         "Reply",
-			CallbackData: fmt.Sprintf("r|%s|%s", receiverUUID, replyMessageID),
-		}
-	}
-
-	_, _, err = cb.Message.EditReplyMarkup(b, &gotgbot.EditMessageReplyMarkupOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-				{
-					replyMessageKey,
-					{
-						Text:         "Block",
-						CallbackData: fmt.Sprintf("b|%s|%s", receiverUUID, replyMessageID),
-					},
-				},
-			},
-		},
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to update message markup: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) unBlockAll(b *gotgbot.Bot, ctx *ext.Context) error {
-	err := r.userRepo.updateBlacklist(r.user.UUID, "clear", "")
-	if err != nil {
-		return fmt.Errorf("failed to unblock all users: %w", err)
-	}
-
-	_, err = ctx.EffectiveMessage.Reply(b, "All users unblocked!", nil)
-	if err != nil {
-		return fmt.Errorf("failed to send bot info: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) manageUsername(b *gotgbot.Bot, ctx *ext.Context) error {
-	var text string
-	var buttons [][]gotgbot.InlineKeyboardButton
-
-	if r.user.Username != "" {
-		text = fmt.Sprintf("Your current username is: %s", r.user.Username)
-		buttons = [][]gotgbot.InlineKeyboardButton{
-			{
-				{
-					Text:         "Change",
-					CallbackData: "u",
-				},
-				{
-					Text:         "Remove",
-					CallbackData: "ru",
-				},
-				{
-					Text:         "Cancel",
-					CallbackData: "cu",
-				},
-			},
-		}
-	} else {
-		text = "You don't have a username!"
-		buttons = [][]gotgbot.InlineKeyboardButton{
-			{
-				{
-					Text:         "Set one",
-					CallbackData: "u",
-				},
-				{
-					Text:         "Cancel",
-					CallbackData: "cu",
-				},
-			},
-		}
-	}
-
-	_, err := b.SendMessage(ctx.EffectiveChat.Id, text, &gotgbot.SendMessageOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: buttons,
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("failed to send username info: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) usernameCallback(b *gotgbot.Bot, ctx *ext.Context, action string) error {
-	cb := ctx.Update.CallbackQuery
-
-	// Remove username command buttons
-	_, _, err := cb.Message.EditReplyMarkup(b, &gotgbot.EditMessageReplyMarkupOpts{})
-	if err != nil {
-		return fmt.Errorf("failed to update username message markup: %w", err)
-	}
-
-	if action == "CANCEL" {
-		// Send callback answer to telegram
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text: "Never mind!",
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-	} else if action == "SET" {
-		err := r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-			"State":          SettingUsername,
-			"ContactUUID":    nil,
-			"ReplyMessageID": nil,
-		})
-		if err != nil {
-			return fmt.Errorf("failed to update user state: %w", err)
-		}
-
-		// Send reply instruction
-		_, err = ctx.EffectiveMessage.Reply(b, "Create a username that starts with a letter, includes 3-20 characters, and may contain letters, numbers, or underscores (_). Usernames are automatically converted to lowercase. \n\nEnter new username:", nil)
-		if err != nil {
-			return fmt.Errorf("failed to send reply message: %w", err)
-		}
-
-		// Send callback answer to telegram
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text: "Setting username...",
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-	} else if action == "REMOVE" {
-		err := r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-			"State":          Idle,
-			"Username":       nil,
-			"ContactUUID":    nil,
-			"ReplyMessageID": nil,
-		})
-		if err != nil {
-			return fmt.Errorf("failed to remove username: %w", err)
-		}
-
-		_, _, err = cb.Message.EditText(b, "Username has been removed!", &gotgbot.EditMessageTextOpts{})
-		if err != nil {
-			return fmt.Errorf("failed to update username message text: %w", err)
-		}
-
-		// Send callback answer to telegram
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text: "Username removed!",
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func (r *RootHandler) setUsername(b *gotgbot.Bot, ctx *ext.Context) error {
-	username := ctx.EffectiveMessage.Text
-
-	if isValidUsername(username) == false {
-		// Send username instruction
-		_, err := ctx.EffectiveMessage.Reply(b, "The entered username is not valid. Enter another one:", nil)
-		if err != nil {
-			return fmt.Errorf("failed to send reply message: %w", err)
-		}
-		return nil
-	}
-
-	// Convert to lowercase
-	username = strings.ToLower(username)
-
-	existingUser, err := r.userRepo.readUserByUsername(username)
-	if err != nil || existingUser == nil {
-		err := r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-			"Username":       username,
-			"State":          Idle,
-			"ContactUUID":    nil,
-			"ReplyMessageID": nil,
-		})
-		if err != nil {
-			return fmt.Errorf("failed to update username: %w", err)
-		}
-
-		// Send username instruction
-		_, err = ctx.EffectiveMessage.Reply(b, fmt.Sprintf("Username has been set: %s", username), nil)
-		if err != nil {
-			return fmt.Errorf("failed to send reply message: %w", err)
-		}
-	} else {
-		var text string
-		if existingUser.UUID != r.user.UUID {
-			text = "The entered username exists. Enter another one:"
-		} else {
-			text = "You already own this username silly! If you want to change it, run the username command once more!"
-
-			// Reset sender user
-			err = r.userRepo.resetUserState(r.user.UUID)
-			if err != nil {
-				return err
-			}
-		}
-		// Send username instruction
-		_, err = ctx.EffectiveMessage.Reply(b, text, nil)
-		if err != nil {
-			return fmt.Errorf("failed to send reply message: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func isValidUsername(username string) bool {
-	// Check length
-	if len(username) < 3 || len(username) > 20 {
-		return false
-	}
-
-	// Regular expression to check valid characters
-	// ^[a-zA-Z0-9_]+$
-	// This checks the string consists only of English letters, digits, and underscores
-	re := regexp.MustCompile(`^[a-zA-Z0-9_]+$`)
-	if !re.MatchString(username) {
-		return false
-	}
-
-	// Check first character (not a digit or underscore)
-	firstChar := username[0]
-	if firstChar == '_' || ('0' <= firstChar && firstChar <= '9') {
-		return false
-	}
-
-	return true
-}
-
-func (r *RootHandler) manageLanguage(b *gotgbot.Bot, ctx *ext.Context) error {
-	var text string
-	if r.user.Language != "" {
-		text = fmt.Sprintf("Your language is: %s", r.user.Language)
-	} else {
-		text = "You don't have a preferred language yet."
-	}
-	_, err := b.SendMessage(ctx.EffectiveChat.Id, text, &gotgbot.SendMessageOpts{
-		ReplyMarkup: gotgbot.InlineKeyboardMarkup{
-			InlineKeyboard: [][]gotgbot.InlineKeyboardButton{
-				{
-					{
-						Text:         "English",
-						CallbackData: fmt.Sprintf("l|%s", i18n.EnUS),
-					},
-					{
-						Text:         "Farsi",
-						CallbackData: fmt.Sprintf("l|%s", i18n.FaIR),
-					},
-					{
-						Text:         "Cancel",
-						CallbackData: "lc",
-					},
-				},
-			},
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("failed to send lkanguage info: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RootHandler) languageCallback(b *gotgbot.Bot, ctx *ext.Context, action string) error {
-	cb := ctx.Update.CallbackQuery
-
-	// Remove language command buttons
-	_, _, err := cb.Message.EditReplyMarkup(b, &gotgbot.EditMessageReplyMarkupOpts{})
-	if err != nil {
-		return fmt.Errorf("failed to update language message markup: %w", err)
-	}
-
-	if action == "CANCEL" {
-		// Send callback answer to telegram
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text: "Never mind!",
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-	} else if action == "SET" {
-		split := strings.Split(cb.Data, "|")
-		if len(split) != 2 {
-			return fmt.Errorf("invalid callback data: %s", cb.Data)
-		}
-
-		var isLanguageValid bool = false
-		for _, lang := range []i18n.Language{i18n.EnUS, i18n.FaIR} {
-			if split[1] == string(lang) {
-				isLanguageValid = true
-				break
-			}
-		}
-
-		if isLanguageValid == false {
-			return fmt.Errorf("invalid language code in callback data: %s", cb.Data)
-		}
-
-		err := r.userRepo.updateUser(r.user.UUID, map[string]interface{}{
-			"State":          Idle,
-			"ContactUUID":    nil,
-			"ReplyMessageID": nil,
-			"Language":       split[1],
-		})
-		if err != nil {
-			return fmt.Errorf("failed to update user language: %w", err)
-		}
-
-		// Send update status
-		_, err = ctx.EffectiveMessage.Reply(b, fmt.Sprintf("Language updated successfully to %s", split[1]), nil)
-		if err != nil {
-			return fmt.Errorf("failed to send language update message: %w", err)
-		}
-
-		// Send callback answer to telegram
-		_, err = cb.Answer(b, &gotgbot.AnswerCallbackQueryOpts{
-			Text: "Language updated!",
-		})
-		if err != nil {
-			return fmt.Errorf("failed to answer callback: %w", err)
-		}
-	}
-	return nil
-}
-
-func blockCheck(sender *User, receiver *User) BlockedBy {
-	if slices.Contains(sender.Blacklist, receiver.UUID) {
-		return Sender
-	} else if slices.Contains(receiver.Blacklist, sender.UUID) {
-		return Receiver
-	}
-	return None
 }
